requirements.txt
httpx>=0.27,<1
supabase>=2,<3
google-generativeai>=0.8,<1
pyyaml>=6,<7
pydantic>=2,<3
